Ship the py.typed marker with the package
This is used to declare that the package contains type information
and that the information can be used by the tools when the package is
installed. See https://www.python.org/dev/peps/pep-0561/ for details.

Unfortunately this requires switching from a module to a package, as
with a module there's nowhere to put the py.typed marker.